Rename init-kafka to kafka-init
