+++ kafka-init/Dockerfile
@@ -3,6 +3,6 @@ FROM strimzi/java-base:latest
 ARG strimzi_version=1.0-SNAPSHOT
 ENV STRIMZI_VERSION ${strimzi_version}
 
-ADD target/init-kafka-${strimzi_version}.jar /
+ADD target/kafka-init-${strimzi_version}.jar /
 
-CMD /bin/launch_java.sh /init-kafka-${STRIMZI_VERSION}.jar
+CMD /bin/launch_java.sh /kafka-init-${STRIMZI_VERSION}.jar
