+++ pom.xml
@@ -42,7 +42,7 @@
         <module>api</module>
         <module>topic-operator</module>
         <module>cluster-operator</module>
-        <module>init-kafka</module>
+        <module>kafka-init</module>
         <module>certificate-manager</module>
         <module>systemtest</module>
     </modules>
